backport: catch up on the latest changes from v0.10 (#528)
* chore(types)!: rename genesis.json quorum fields (#515)

* fix(consensus): ProcessProposal executed twice for a block (#516)

* fix(consensus): ProcessProposal executed twice for a block

* fix(consensus): ensure process proposal is called once before finalize block

* fix(consensus): ensure ProcessProposal is called at least once for a block

* build(deps): Bump actions/setup-go from 3.3.1 to 3.4.0 (#524)

Bumps [actions/setup-go](https://github.com/actions/setup-go) from 3.3.1 to 3.4.0.
- [Release notes](https://github.com/actions/setup-go/releases)
- [Commits](https://github.com/actions/setup-go/compare/v3.3.1...v3.4.0)

---
updated-dependencies:
- dependency-name: actions/setup-go
  dependency-type: direct:production
  update-type: version-update:semver-minor
...



* build(deps): Bump bufbuild/buf-setup-action from 1.9.0 to 1.10.0 (#525)

Bumps [bufbuild/buf-setup-action](https://github.com/bufbuild/buf-setup-action) from 1.9.0 to 1.10.0.
- [Release notes](https://github.com/bufbuild/buf-setup-action/releases)
- [Commits](https://github.com/bufbuild/buf-setup-action/compare/v1.9.0...v1.10.0)

---
updated-dependencies:
- dependency-name: bufbuild/buf-setup-action
  dependency-type: direct:production
  update-type: version-update:semver-minor
...



* fix(consensus): proposer-based timestamp broken during backport (#523)

* fix(consensus): proposer-based timestamp is broken

* fix(consensus): initial block time  must be equal genesis time

* test(consensus): update TestStateProposalTime for PBTS

* test(consensus): fix TestStateLock_POLSafety2 for PBTS

* Revert "test(consensus): skip broken PBTS tests (#500)"

This reverts commit 02452f199625ce4fdbaaadca50504182f51b3c44.

* refactor(time): simplify time processing

* feat(log): Allow definition of log assertions in TestingLogger

* test(consensus): refactor PBTS tests

* chore(log): replace TestingLogger interface with struct

* chore(state): move block time logic to separate fn

* chore(abci)!: remove Snapshot.core_chain_locked_height (#527)

* chore(abci)!: remove Snapshot.core_chain_locked_height

* build: bump Golang to 1.19

* fix(wal): improve wal replay mechanism (#510)

* refactor: improve wal iterator

* feat: add config parameter to turn on/off skipping the WAL rounds to the last

* refactor: clean up code

* refactor: add TestWALRoundsSkipper

* fix: derive a commit from vote-set during replay

* fix: add validators priority incrementing to select proper proposer

* chore(release): update changelog and version to 0.10.0-dev.6 (#526)

+++ internal/consensus/wal_generator_test.go
@@ -21,14 +21,14 @@ import (
 // consensus state) with a kvstore application and special consensus
 // wal instance (byteBufferWAL) and waits until numBlocks are created.
 // If the node fails to produce given numBlocks, it fails the test.
-func WALGenerateNBlocks(ctx context.Context, t *testing.T, logger log.Logger, wr io.Writer, node *fakeNode, numBlocks int) {
+func WALGenerateNBlocks(ctx context.Context, t *testing.T, logger log.Logger, wr io.Writer, node *fakeNode, numBlocks int64) {
 	t.Helper()
 
 	logger.Info("generating WAL (last height msg excluded)", "numBlocks", numBlocks)
 
 	// set consensus wal to buffered WAL, which will write all incoming msgs to buffer
 	numBlocksWritten := make(chan struct{})
-	wal := newByteBufferWAL(logger, NewWALEncoder(wr), int64(numBlocks), numBlocksWritten)
+	wal := newByteBufferWAL(logger, NewWALEncoder(wr), numBlocks, numBlocksWritten)
 	// see wal.go#103
 	if err := wal.Write(EndHeightMessage{0}); err != nil {
 		t.Fatal(err)
@@ -50,7 +50,7 @@ func WALGenerateNBlocks(ctx context.Context, t *testing.T, logger log.Logger, wr
 }
 
 // WALWithNBlocks returns a WAL content with numBlocks.
-func WALWithNBlocks(ctx context.Context, t *testing.T, logger log.Logger, node *fakeNode, numBlocks int) (data []byte, err error) {
+func WALWithNBlocks(ctx context.Context, t *testing.T, logger log.Logger, node *fakeNode, numBlocks int64) (data []byte, err error) {
 	var b bytes.Buffer
 	wr := bufio.NewWriter(&b)
